requirements.txt
attrs==20.2.0
aws-cdk.aws-iam==1.61.1
aws-cdk.cloud-assembly-schema==1.61.1
aws-cdk.core==1.61.1
aws-cdk.cx-api==1.61.1
aws-cdk.region-info==1.61.1
## !! Could not determine repository location
brev-iam-maker==0.0.1
cattrs==1.0.0
constructs==3.0.4
jsii==1.12.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
